requirements.txt
Django>=4.2
scikit-learn
pandas
numpy
matplotlib
seaborn
joblib
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-database-url==2.3.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
load-dotenv==0.1.0
multidict==6.3.2
mysqlclient==2.2.7
openpyxl==3.1.5
packaging==24.2
propcache==0.3.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
twilio==9.5.1
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
yarl==1.18.3
